+++ src/ContainerObject/class.ContainerObjectsCtrl.php
@@ -86,7 +86,7 @@ class ContainerObjectsCtrl
     {
         $count = self::srContainerObjectMenu()->menu()->cleanUpLostItems();
 
-        ilUtil::sendInfo(self::plugin()->translate("cleaned_up_lost_items",self::LANG_MODULE,[
+        ilUtil::sendInfo(self::plugin()->translate("cleaned_up_lost_items", self::LANG_MODULE, [
             $count
         ]), true);